azur: avoid -L interface flag that would expose libraries to -l
Now that all the libraries are in lib/azur, we switched to proper names,
libimgui.a, libavcodec.a, etc. Azur::Azur's `-L ${AZUR_LIB}` would
expose these to -limgui, -lavcodec which is unexpected and unlikely to
be useful. It's also unneeded since CMake links with the path to the .a
file on the command-line directly.